requirement.txt
alembic==1.2.1
aniso8601==7.0.0
asgiref==3.2.2
boto3==1.10.0
botocore==1.13.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
h11==0.8.1
httptools==0.0.13
hupper==1.9
idna==2.8
Jinja2==2.10.3
jmespath==0.9.4
Mako==1.1.0
MarkupSafe==1.1.1
-e git+https://github.com/KAMESTERY/middlewarekam.git@0e3b8f3ecae1699309a665f2bb261432aa510a95#egg=middlewarekam
PasteDeploy==2.0.1
plaster==1.0
plaster-pastedeploy==0.7
promise==2.2.1
Pygments==2.4.2
pyramid==1.10.4
pyramid-debugtoolbar==4.5
pyramid-jinja2==2.8
pyramid-mako==1.1.0
pyramid-retry==2.1
pyramid-tm==2.3
python-dateutil==2.8.0
python-editor==1.0.4
repoze.lru==0.7
requests==2.22.0
Rx==1.6.1
s3transfer==0.2.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.10
transaction==2.4.0
translationstring==1.3
urllib3==1.25.6
uvicorn==0.9.0
uvloop==0.13.0
venusian==1.2.0
waitress==1.3.1
WebOb==1.8.5
websockets==8.0.2
youtube-dl==2019.10.22
zope.deprecation==4.4.0
zope.interface==4.6.0
zope.sqlalchemy==1.2
